Fixes #456 - Our data directory should adopt standard python package structure
